fix: 삭제 할 때 삭제 확인 창 뜨게 설정 - 피드삭제, 댓글삭제, 물품삭제, 회원탈퇴

+++ frontend/src/components/group/Comment.vue
@@ -252,14 +252,17 @@ export default {
         })
     },
     deleteComment(commentPk){
-      http.delete(`comment/deleteComment?commentPk=${commentPk}`)
+      if(confirm("댓글을 삭제하시겠습니까?")==true) {
+        http.delete(`comment/deleteComment?commentPk=${commentPk}`)
         .then((res)=>{
           if(res.data.status){
+            alert("댓글이 삭제되었습니다.")
             this.getComment()
           }
         }).catch((err) => {
           console.log(err)
         })
+      }         
     },
     updateComment(comment) {
       const params = {
